+++ test/variational/test_variational_mixed.py
@@ -106,17 +106,17 @@ def test_n_samples_api(vstate, _mpi_size):
     vstate.n_samples = 3
     check_consistent(vstate, _mpi_size)
     assert vstate.samples.shape[0:2] == (
-        int(np.ceil(3 / _mpi_size)),
         vstate.sampler.n_chains_per_rank,
+        int(np.ceil(3 / _mpi_size)),
     )
 
     vstate.n_samples_per_rank = 4
     check_consistent(vstate, _mpi_size)
-    assert vstate.samples.shape[0:2] == (4, vstate.sampler.n_chains_per_rank)
+    assert vstate.samples.shape[0:2] == (vstate.sampler.n_chains_per_rank, 4)
 
     vstate.chain_length = 2
     check_consistent(vstate, _mpi_size)
-    assert vstate.samples.shape[0:2] == (2, vstate.sampler.n_chains_per_rank)
+    assert vstate.samples.shape[0:2] == (vstate.sampler.n_chains_per_rank, 2)
 
     vstate.n_samples = 1000
     vstate.n_discard_per_chain = None
@@ -137,16 +137,16 @@ def test_n_samples_api(vstate, _mpi_size):
     vstate.n_samples = 3
     check_consistent(vstate, _mpi_size)
     # `n_samples` is rounded up
-    assert vstate.samples.shape[0:2] == (1, vstate.sampler.n_chains_per_rank)
+    assert vstate.samples.shape[0:2] == (vstate.sampler.n_chains_per_rank, 1)
 
     vstate.n_samples_per_rank = 16 // _mpi_size + 1
     check_consistent(vstate, _mpi_size)
     # `n_samples` is rounded up
-    assert vstate.samples.shape[0:2] == (2, vstate.sampler.n_chains_per_rank)
+    assert vstate.samples.shape[0:2] == (vstate.sampler.n_chains_per_rank, 2)
 
     vstate.chain_length = 2
     check_consistent(vstate, _mpi_size)
-    assert vstate.samples.shape[0:2] == (2, vstate.sampler.n_chains_per_rank)
+    assert vstate.samples.shape[0:2] == (vstate.sampler.n_chains_per_rank, 2)
 
 
 def test_n_samples_diag_api(vstate, _mpi_size):
@@ -170,16 +170,16 @@ def test_n_samples_diag_api(vstate, _mpi_size):
     check_consistent_diag(vstate)
     assert (
         vstate.diagonal.samples.shape[0:2]
-        == (int(np.ceil(3 / _mpi_size)), vstate.sampler_diag.n_chains_per_rank)
-        == (int(np.ceil(3 / _mpi_size)), vstate.diagonal.sampler.n_chains_per_rank)
+        == (vstate.sampler_diag.n_chains_per_rank, int(np.ceil(3 / _mpi_size)))
+        == (vstate.diagonal.sampler.n_chains_per_rank, int(np.ceil(3 / _mpi_size)))
     )
 
     vstate.chain_length_diag = 2
     check_consistent_diag(vstate)
     assert (
         vstate.diagonal.samples.shape[0:2]
-        == (2, vstate.sampler_diag.n_chains_per_rank)
-        == (2, vstate.diagonal.sampler.n_chains_per_rank)
+        == (vstate.sampler_diag.n_chains_per_rank, 2)
+        == (vstate.diagonal.sampler.n_chains_per_rank, 2)
     )
 
     vstate.n_samples_diag = 1000
@@ -205,8 +205,8 @@ def test_n_samples_diag_api(vstate, _mpi_size):
     # `n_samples_diag` is rounded up
     assert (
         vstate.diagonal.samples.shape[0:2]
-        == (1, vstate.sampler_diag.n_chains_per_rank)
-        == (1, vstate.diagonal.sampler.n_chains_per_rank)
+        == (vstate.sampler_diag.n_chains_per_rank, 1)
+        == (vstate.diagonal.sampler.n_chains_per_rank, 1)
     )
 
     vstate.chain_length_diag = 2
@@ -218,8 +218,8 @@ def test_n_samples_diag_api(vstate, _mpi_size):
     )
     assert (
         vstate.diagonal.samples.shape[0:2]
-        == (2, vstate.sampler_diag.n_chains_per_rank)
-        == (2, vstate.diagonal.sampler.n_chains_per_rank)
+        == (vstate.sampler_diag.n_chains_per_rank, 2)
+        == (vstate.diagonal.sampler.n_chains_per_rank, 2)
     )