+++ CHANGELOG.md
@@ -10,9 +10,11 @@ This release requires Python 3.8 and Jax 0.4.
 * `netket.callbacks.EarlyStopping` now supports relative tolerances for determining when to stop [#14XX](https://github.com/netket/netket/pull/1481).
 * `netket.callbacks.ConvergenceStopping` has been added, which can stop a driver when the loss function reaches a certain threshold [#14XX](https://github.com/netket/netket/pull/1481).
 
+### Breaking Changes
+* The first two axes in the output of the samplers have been swapped, samples are now of shape `(n_chains, n_samples_per_chain, ...)` consistent with `netket.stats.statistics`. Custom samplers need to be updated to return arrays of shape `(n_chains, n_samples_per_chain, ...)` instead of `(n_samples_per_chain, n_chains, ...)`. [#1502](https://github.com/netket/netket/pull/1502)
+
 ### Deprecations
 * `netket.vqs.ExactState` has been renamed to {class}`netket.vqs.FullSumState` to better reflect what it does. Using the old name will now raise a warning [#1477](https://github.com/netket/netket/pull/1477).
-
 
 
 ## NetKet 3.8 (8 May 2023)